Extracted out findClosestStart
